Revert "Merge branch 'CC-58' into CC-81"
This reverts commit cc1690cf938ef032186d5891b1e1c9e7189c4a9d.

+++ CoCoCoWorking.DAL/DTO/OrderDTO.cs
@@ -10,7 +10,7 @@ namespace CoCoCoWorking.DAL.DTO
     {
         public int Id { get; set; }
         public int CustomerId { get; set; }
-        public decimal? OrderCost { get; set; }
+        public int? OrderCost { get; set; }
         public string OrderStatus { get; set; }
         public string PaidDate { get; set; }
         public List<OrderUnitDto> OrderUnits { get; set; } = new List<OrderUnitDto>();
